+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     - also updated via ``LineEditor#readLine`` method
     - when receive ``SIGWINCH``, automatically updated
 - in ``LineEditor#readLine``, scroll lines when input rows is larger than window rows
+- add ``String#ifEmpty`` method
 
 #### Misc